Revert "WIP: Refactor includes to follow <boost/gil/...hpp> convention"
