+++ include/boost/gil.hpp
@@ -5,28 +5,27 @@
 // (See accompanying file LICENSE_1_0.txt
 // or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef BOOST_GIL_GIL_HPP
-#define BOOST_GIL_GIL_HPP
+#ifndef boost_gil_hpp_
+#define boost_gil_hpp_
 
-#include <boost/gil/config.hpp>
 #include <boost/gil/version.hpp>
-
-#include <boost/gil/algorithm.hpp>
-#include <boost/gil/bit_aligned_pixel_iterator.hpp>
+#include <boost/gil/gil_config.hpp>
 #include <boost/gil/channel_algorithm.hpp>
+#include <boost/gil/algorithm.hpp>
+#include <boost/gil/pixel.hpp>
+#include <boost/gil/packed_pixel.hpp>
+#include <boost/gil/planar_pixel_reference.hpp>
+#include <boost/gil/planar_pixel_iterator.hpp>
+#include <boost/gil/pixel_iterator_adaptor.hpp>
+#include <boost/gil/step_iterator.hpp>
+#include <boost/gil/iterator_from_2d.hpp>
+#include <boost/gil/image.hpp>
+#include <boost/gil/image_view_factory.hpp>
+#include <boost/gil/typedefs.hpp>
+#include <boost/gil/metafunctions.hpp>
 #include <boost/gil/color_convert.hpp>
 #include <boost/gil/device_n.hpp>
-#include <boost/gil/image_view_factory.hpp>
-#include <boost/gil/image.hpp>
-#include <boost/gil/iterator_from_2d.hpp>
-#include <boost/gil/metafunctions.hpp>
-#include <boost/gil/packed_pixel.hpp>
-#include <boost/gil/pixel_iterator_adaptor.hpp>
-#include <boost/gil/pixel.hpp>
-#include <boost/gil/planar_pixel_iterator.hpp>
-#include <boost/gil/planar_pixel_reference.hpp>
-#include <boost/gil/step_iterator.hpp>
-#include <boost/gil/typedefs.hpp>
 #include <boost/gil/virtual_locator.hpp>
+#include <boost/gil/bit_aligned_pixel_iterator.hpp>
 
 #endif
