+++ include/boost/gil/deprecated.hpp
@@ -10,8 +10,8 @@
 
 /*************************************************************************************************/
 
-#ifndef BOOST_GIL_DEPRECATED_HPP
-#define BOOST_GIL_DEPRECATED_HPP
+#ifndef GIL_DEPRECATED_HPP
+#define GIL_DEPRECATED_HPP
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file
