+++ include/boost/gil/gil_concept.hpp
@@ -10,8 +10,8 @@
 
 /*************************************************************************************************/
 
-#ifndef BOOST_GIL_CONCEPT_HPP
-#define BOOST_GIL_CONCEPT_HPP
+#ifndef GIL_CONCEPT_H
+#define GIL_CONCEPT_H
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file               
@@ -22,16 +22,14 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <boost/gil/config.hpp>
-
+#include <functional>
+#include "gil_config.hpp"
+#include <boost/type_traits.hpp>
+#include <boost/utility/enable_if.hpp>
 #include <boost/concept_check.hpp>
 #include <boost/iterator/iterator_concepts.hpp>
 #include <boost/mpl/and.hpp>
 #include <boost/mpl/size.hpp>
-#include <boost/type_traits.hpp>
-#include <boost/utility/enable_if.hpp>
-
-#include <functional>
 
 namespace boost { namespace gil {
