+++ include/boost/gil/packed_pixel.hpp
@@ -10,8 +10,8 @@
 
 /*************************************************************************************************/
 
-#ifndef BOOST_GIL_PACKED_PIXEL_HPP
-#define BOOST_GIL_PACKED_PIXEL_HPP
+#ifndef GIL_PACKED_PIXEL_H
+#define GIL_PACKED_PIXEL_H
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file               
@@ -22,15 +22,13 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <boost/gil/config.hpp>
-#include <boost/gil/pixel.hpp>
-
+#include <functional>
 #include <boost/core/ignore_unused.hpp>
+#include <boost/utility/enable_if.hpp>
 #include <boost/mpl/bool.hpp>
 #include <boost/mpl/front.hpp>
-#include <boost/utility/enable_if.hpp>
-
-#include <functional>
+#include "gil_config.hpp"
+#include "pixel.hpp"
 
 namespace boost { namespace gil {
