+++ include/boost/gil/pixel.hpp
@@ -9,8 +9,8 @@
 */
 /*************************************************************************************************/
 
-#ifndef BOOST_GIL_PIXEL_HPP
-#define BOOST_GIL_PIXEL_HPP
+#ifndef GIL_PIXEL_H
+#define GIL_PIXEL_H
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// \file               
@@ -21,21 +21,19 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////////////
 
-#include <boost/gil/config.hpp>
-#include <boost/gil/channel.hpp>
-#include <boost/gil/color_base_algorithm.hpp>
-#include <boost/gil/color_base.hpp>
-#include <boost/gil/concept.hpp>
-#include <boost/gil/metafunctions.hpp>
-#include <boost/gil/utility.hpp>
-
+#include <functional>
 #include <boost/core/ignore_unused.hpp>
 #include <boost/utility/enable_if.hpp>
 #include <boost/mpl/bool.hpp>
 #include <boost/mpl/front.hpp>
 #include <boost/type_traits.hpp>
-
-#include <functional>
+#include "gil_config.hpp"
+#include "color_base.hpp"
+#include "gil_concept.hpp"
+#include "channel.hpp"
+#include "metafunctions.hpp"
+#include "utilities.hpp"
+#include "color_base_algorithm.hpp"
 
 namespace boost { namespace gil {
